Add personal settings frame and form
The frame shows when the user clicks on
the personal settings button in the nav bar.
The form's placeholders are prepopulated with
the current data associated with that particular
users' data as stored in the database. Likewise,
updates made to the input fields are stored in the
database if the user presses the submit button.