Refactor exercise-selection frame
Added a conditional to the post workout cleanup in
the ui module such that if a new workout is
created but the user exits the process without
creating any exercises, this workout is also
deleted from the database as it has zero content.

The search checkbox in the same frame is now
automatically set to unchecked upon showing this
frame.