Refactor code
Added spin.js - a lightweight spinner which has
been added to the sign-in and sign-up features.

Added a page-reset feature to the resolved promise
handler of the sign-out request. This includes the
resetting of all form elements as well as ensuring
the navbar is collapsed if the same page is used
for another sign-in.